Move more test utils
