+++ crates/blockifier/src/transaction/transactions_test.rs
@@ -32,69 +32,23 @@ use crate::state::cached_state::CachedState;
 use crate::state::errors::StateError;
 use crate::state::state_api::{State, StateReader};
 use crate::test_utils::{
-    test_erc20_account_balance_key, test_erc20_faulty_account_balance_key,
-    test_erc20_sequencer_balance_key, DictStateReader, ACCOUNT_CONTRACT_PATH, BALANCE,
-    ERC20_CONTRACT_PATH, MAX_FEE, TEST_ACCOUNT_CONTRACT_ADDRESS, TEST_ACCOUNT_CONTRACT_CLASS_HASH,
-    TEST_CLASS_HASH, TEST_CONTRACT_ADDRESS, TEST_CONTRACT_PATH, TEST_EMPTY_CONTRACT_CLASS_HASH,
-    TEST_EMPTY_CONTRACT_PATH, TEST_ERC20_CONTRACT_ADDRESS, TEST_ERC20_CONTRACT_CLASS_HASH,
+    test_erc20_account_balance_key, test_erc20_sequencer_balance_key, DictStateReader, BALANCE,
+    MAX_FEE, TEST_ACCOUNT_CONTRACT_ADDRESS, TEST_ACCOUNT_CONTRACT_CLASS_HASH, TEST_CLASS_HASH,
+    TEST_CONTRACT_ADDRESS, TEST_EMPTY_CONTRACT_CLASS_HASH, TEST_EMPTY_CONTRACT_PATH,
+    TEST_ERC20_CONTRACT_ADDRESS, TEST_ERC20_CONTRACT_CLASS_HASH,
     TEST_FAULTY_ACCOUNT_CONTRACT_ADDRESS, TEST_FAULTY_ACCOUNT_CONTRACT_CLASS_HASH,
-    TEST_FAULTY_ACCOUNT_CONTRACT_PATH,
 };
 use crate::transaction::account_transaction::AccountTransaction;
 use crate::transaction::constants;
 use crate::transaction::errors::TransactionExecutionError;
 use crate::transaction::objects::{ResourcesMapping, TransactionExecutionInfo};
-use crate::transaction::test_utils::create_account_tx_for_validate_test;
+use crate::transaction::test_utils::{
+    create_account_tx_for_validate_test, create_state_with_falliable_validation_account,
+    create_state_with_trivial_validation_account, CALL_CONTRACT, INVALID, VALID,
+};
 use crate::transaction::transaction_execution::Transaction;
 use crate::transaction::transaction_types::TransactionType;
 use crate::transaction::transactions::{DeclareTransaction, ExecutableTransaction};
-
-// Corresponding constants to the ones in faulty_account.
-pub const VALID: u64 = 0;
-pub const INVALID: u64 = 1;
-pub const CALL_CONTRACT: u64 = 2;
-
-fn create_account_tx_test_state(
-    account_class_hash: &str,
-    account_address: &str,
-    account_path: &str,
-    erc20_account_balance_key: StorageKey,
-    initial_account_balance: u128,
-) -> CachedState<DictStateReader> {
-    let block_context = BlockContext::create_for_testing();
-
-    let test_contract_class_hash = ClassHash(stark_felt!(TEST_CLASS_HASH));
-    let test_account_class_hash = ClassHash(stark_felt!(account_class_hash));
-    let test_erc20_class_hash = ClassHash(stark_felt!(TEST_ERC20_CONTRACT_CLASS_HASH));
-    let class_hash_to_class = HashMap::from([
-        (test_account_class_hash, ContractClassV0::from_file(account_path).into()),
-        (test_contract_class_hash, ContractClassV0::from_file(TEST_CONTRACT_PATH).into()),
-        (test_erc20_class_hash, ContractClassV0::from_file(ERC20_CONTRACT_PATH).into()),
-    ]);
-    let test_contract_address = ContractAddress(patricia_key!(TEST_CONTRACT_ADDRESS));
-    // A random address that is unlikely to equal the result of the calculation of a contract
-    // address.
-    let test_account_address = ContractAddress(patricia_key!(account_address));
-    let test_erc20_address = block_context.fee_token_address;
-    let address_to_class_hash = HashMap::from([
-        (test_contract_address, test_contract_class_hash),
-        (test_account_address, test_account_class_hash),
-        (test_erc20_address, test_erc20_class_hash),
-    ]);
-    let minter_var_address = get_storage_var_address("permitted_minter", &[])
-        .expect("Failed to get permitted_minter storage address.");
-    let storage_view = HashMap::from([
-        ((test_erc20_address, erc20_account_balance_key), stark_felt!(initial_account_balance)),
-        // Give the account mint permission.
-        ((test_erc20_address, minter_var_address), *test_account_address.0.key()),
-    ]);
-    CachedState::new(DictStateReader {
-        address_to_class_hash,
-        class_hash_to_class,
-        storage_view,
-        ..Default::default()
-    })
-}
 
 fn expected_validate_call_info(
     class_hash: ClassHash,
@@ -252,28 +206,6 @@ fn invoke_tx() -> InvokeTransactionV1 {
         ContractAddress(patricia_key!(TEST_ACCOUNT_CONTRACT_ADDRESS)),
         Fee(MAX_FEE),
         None,
-    )
-}
-
-fn create_state_with_trivial_validation_account() -> CachedState<DictStateReader> {
-    let account_balance = BALANCE;
-    create_account_tx_test_state(
-        TEST_ACCOUNT_CONTRACT_CLASS_HASH,
-        TEST_ACCOUNT_CONTRACT_ADDRESS,
-        ACCOUNT_CONTRACT_PATH,
-        test_erc20_account_balance_key(),
-        account_balance,
-    )
-}
-
-fn create_state_with_falliable_validation_account() -> CachedState<DictStateReader> {
-    let account_balance = BALANCE;
-    create_account_tx_test_state(
-        TEST_FAULTY_ACCOUNT_CONTRACT_CLASS_HASH,
-        TEST_FAULTY_ACCOUNT_CONTRACT_ADDRESS,
-        TEST_FAULTY_ACCOUNT_CONTRACT_PATH,
-        test_erc20_faulty_account_balance_key(),
-        account_balance * 2,
     )
 }
